Improve scripts to run coupled GEFS workflow
 On branch feature/gefs_v13_coupled
 Changes to be committed:
	modified:   jobs/JGLOBAL_WAVE_INIT
	modified:   jobs/rocoto/coupled_ic.sh
	modified:   jobs/rocoto/waveinit.sh
	modified:   parm/config/config.base.emc.dyn
	modified:   parm/config/config.coupled_ic
	modified:   parm/config/config.nsst
	modified:   parm/config/config.wave
	modified:   ush/forecast_postdet.sh
	modified:   ush/forecast_predet.sh
	modified:   workflow/applications.py
	modified:   workflow/hosts/hera.yaml
	modified:   workflow/rocoto/workflow_tasks.py
	modified:   workflow/rocoto/workflow_xml.py

+++ jobs/rocoto/coupled_ic.sh
@@ -21,6 +21,9 @@ status=$?
 [[ $status -ne 0 ]] && exit $status
 err=0
 
+if [[ $CDUMP == "gefs" ]]; then
+    export RUNMEM=${RUNMEM:-"c00"}
+fi
 ###############################################################
 # Source relevant configs
 configs="base coupled_ic wave"
@@ -37,10 +40,18 @@ status=$?
 [[ $status -ne 0 ]] && exit $status
 
 # Create ICSDIR if needed
-[[ ! -d $ICSDIR/$CDATE ]] && mkdir -p $ICSDIR/$CDATE
-[[ ! -d $ICSDIR/$CDATE/atmos ]] && mkdir -p $ICSDIR/$CDATE/atmos
-[[ ! -d $ICSDIR/$CDATE/ocn ]] && mkdir -p $ICSDIR/$CDATE/ocn
-[[ ! -d $ICSDIR/$CDATE/ice ]] && mkdir -p $ICSDIR/$CDATE/ice
+if [[ $CDUMP == "gefs" ]]; then
+    ICSDIR=${ROTDIR}/${CDUMP}.${PDY}/${cyc}/${RUNMEM} #/atmos
+    [[ ! -d $ICSDIR ]] && mkdir -p $ICSDIR
+    [[ ! -d $ICSDIR/atmos ]] && mkdir -p $ICSDIR/atmos
+    [[ ! -d $ICSDIR/ocean ]] && mkdir -p $ICSDIR/ocean
+    [[ ! -d $ICSDIR/ice ]] && mkdir -p $ICSDIR/ice
+else
+    [[ ! -d $ICSDIR/$CDATE ]] && mkdir -p $ICSDIR/$CDATE
+    [[ ! -d $ICSDIR/$CDATE/atmos ]] && mkdir -p $ICSDIR/$CDATE/atmos
+    [[ ! -d $ICSDIR/$CDATE/ocn ]] && mkdir -p $ICSDIR/$CDATE/ocn
+    [[ ! -d $ICSDIR/$CDATE/ice ]] && mkdir -p $ICSDIR/$CDATE/ice
+fi
 
 if [ $ICERES = '025' ]; then
   ICERESdec="0.25"
@@ -50,7 +61,11 @@ if [ $ICERES = '050' ]; then
 fi 
 
 # Setup ATM initial condition files
-cp -r $BASE_CPLIC/$CPL_ATMIC/$CDATE/$CDUMP/*  $ICSDIR/$CDATE/atmos/
+if [[ $CDUMP == "gefs" ]]; then
+    cp -r $BASE_CPLIC/$CPL_ATMIC/$CDATE/gfs/$CASE/INPUT $ICSDIR/atmos/
+else
+    cp -r $BASE_CPLIC/$CPL_ATMIC/$CDATE/$CDUMP/*  $ICSDIR/$CDATE/atmos/
+fi
 rc=$?
 if [[ $rc -ne 0 ]] ; then
   echo "FATAL: Unable to copy $BASE_CPLIC/$CPL_ATMIC/$CDATE/$CDUMP/* to $ICSDIR/$CDATE/atmos/ (Error code $rc)" 
@@ -59,15 +74,23 @@ fi
 
 
 # Setup Ocean IC files 
-cp -r $BASE_CPLIC/$CPL_OCNIC/$CDATE/ocn/$OCNRES/MOM*.nc  $ICSDIR/$CDATE/ocn/
+if [[ $CDUMP == "gefs" ]]; then
+    cp -r $BASE_CPLIC/$CPL_OCNIC/$CDATE/ocn/$OCNRES/MOM*.nc  $ICSDIR/ocean/
+else
+    cp -r $BASE_CPLIC/$CPL_OCNIC/$CDATE/ocn/$OCNRES/MOM*.nc  $ICSDIR/$CDATE/ocn/
+fi
 rc=$?
 if [[ $rc -ne 0 ]] ; then
   echo "FATAL: Unable to copy $BASE_CPLIC/$CPL_OCNIC/$CDATE/ocn/$OCNRES/MOM*.nc to $ICSDIR/$CDATE/ocn/ (Error code $rc)"
 fi
 ((err+=$rc))
 
-#Setup Ice IC files 
-cp $BASE_CPLIC/$CPL_ICEIC/$CDATE/ice/$ICERES/cice5_model_${ICERESdec}.res_$CDATE.nc $ICSDIR/$CDATE/ice/cice_model_${ICERESdec}.res_$CDATE.nc
+#Setup Ice IC files
+if [[ $CDUMP == "gefs" ]]; then
+    cp $BASE_CPLIC/$CPL_ICEIC/$CDATE/ice/$ICERES/cice5_model_${ICERESdec}.res_$CDATE.nc $ICSDIR/ice/cice_model_${ICERESdec}.res_$CDATE.nc
+else
+    cp $BASE_CPLIC/$CPL_ICEIC/$CDATE/ice/$ICERES/cice5_model_${ICERESdec}.res_$CDATE.nc $ICSDIR/$CDATE/ice/cice_model_${ICERESdec}.res_$CDATE.nc
+fi
 rc=$?
 if [[ $rc -ne 0 ]] ; then
   echo "FATAL: Unable to copy $BASE_CPLIC/$CPL_ICEIC/$CDATE/ice/$ICERES/cice5_model_${ICERESdec}.res_$CDATE.nc to $ICSDIR/$CDATE/ice/cice_model_${ICERESdec}.res_$CDATE.nc (Error code $rc)"
@@ -75,10 +98,18 @@ fi
 ((err+=$rc))
 
 if [ $DO_WAVE = "YES" ]; then
-  [[ ! -d $ICSDIR/$CDATE/wav ]] && mkdir -p $ICSDIR/$CDATE/wav
+  if [[ $CDUMP == "gefs" ]]; then
+    [[ ! -d $ICSDIR/wave/restart ]] && mkdir -p $ICSDIR/wave/restart/
+  else
+    [[ ! -d $ICSDIR/$CDATE/wav ]] && mkdir -p $ICSDIR/$CDATE/wav
+  fi
   for grdID in $waveGRD
   do
-    cp $BASE_CPLIC/$CPL_WAVIC/$CDATE/wav/$grdID/*restart.$grdID $ICSDIR/$CDATE/wav/
+    if [[ $CDUMP == "gefs" ]]; then
+      cp $BASE_CPLIC/$CPL_WAVIC/$CDATE/wav/$grdID/*restart.$grdID $ICSDIR/wave/restart/
+    else
+      cp $BASE_CPLIC/$CPL_WAVIC/$CDATE/wav/$grdID/*restart.$grdID $ICSDIR/$CDATE/wav/
+    fi
     rc=$?
     if [[ $rc -ne 0 ]] ; then
       echo "FATAL: Unable to copy $BASE_CPLIC/$CPL_WAVIC/$CDATE/wav/$grdID/*restart.$grdID to $ICSDIR/$CDATE/wav/ (Error code $rc)" 
@@ -88,20 +119,24 @@ if [ $DO_WAVE = "YES" ]; then
 fi
 
 # Stage the FV3 initial conditions to ROTDIR
-export OUTDIR="$ICSDIR/$CDATE/atmos/$CASE/INPUT"
-COMOUT="$ROTDIR/$CDUMP.$PDY/$cyc/atmos"
-[[ ! -d $COMOUT ]] && mkdir -p $COMOUT
-cd $COMOUT || exit 99
-rm -rf INPUT
-$NLN $OUTDIR .
+if [[ $CDUMP != "gefs" ]]; then
+    export OUTDIR="$ICSDIR/$CDATE/atmos/$CASE/INPUT"
+    COMOUT="$ROTDIR/$CDUMP.$PDY/$cyc/atmos"
+    [[ ! -d $COMOUT ]] && mkdir -p $COMOUT
+    cd $COMOUT || exit 99
+    rm -rf INPUT
+    $NLN $OUTDIR .
+fi
 
 #Stage the WW3 initial conditions to ROTDIR 
 if [ $DO_WAVE = "YES" ]; then
-  export OUTDIRw="$ICSDIR/$CDATE/wav"
-  COMOUTw="$ROTDIR/$CDUMP.$PDY/$cyc/wave/restart"
-  [[ ! -d $COMOUTw ]] && mkdir -p $COMOUTw
-  cd $COMOUTw || exit 99
-  $NLN $OUTDIRw/* .
+  if [[ $CDUMP != "gefs" ]]; then
+    export OUTDIRw="$ICSDIR/$CDATE/wav"
+    COMOUTw="$ROTDIR/$CDUMP.$PDY/$cyc/wave/restart"
+    [[ ! -d $COMOUTw ]] && mkdir -p $COMOUTw
+    cd $COMOUTw || exit 99
+    $NLN $OUTDIRw/* .
+  fi
 fi
 
 if  [[ $err -ne 0 ]] ; then
